🎁 Allow admin_set_id to be passed in through URL
This commit will let the admin_set_id be passed in through the URL.  The
user can use the collection id to pass in the admin_set_id that they
want to deposit the work into.  If the id is not a valid admin set then
it will fall back to the default admin set.

Also, we want to priortize the Valkyrie resource version of the work so
if the user passes in `--header 'Hyrax-Work-Model: Etd' \` in the
headers, then it will still create an EtdResource.

Ref:
- https://github.com/notch8/palni_palci_knapsack/issues/292