⚙️ Dockerize willow_sword
This commit will allow devs to spin up a Hyrax 5 dassie instance with
WillowSword installed.  The instance will be Valkyrie transition enabled
and built off the latest hyrax-dev image.  We might want to pin it to
something if things get funky. In order to build the solr core, I copied
the solr/ dir from Hyrax's .dassie/solr/ over.  I added an install
generator to more easily get the WillowSword route mounted and generate
migrations.

Also, I think it makes sense to default the authorize_request to `true`.
I feel like it would be a huge security risk to have it set to `false`.