🎁 Add route for works
This commit will add a new route for works which should reach parity
with the existing works route.  In the test it revealed some major
differences between Hyrax and Hyku that makes me want to spin up a Hyku
instance instead.  I think it should be fine for now but I will revisit
that idea when it becomes too painful.