🐛 Move type to a namespace
The `type` element is not a valid in the Atom-Pub schema so we are
moving it into the `h4csys` namespace.  This affects both the service
document and the collection end points.  Since it was a low lift, we are
also backporting the fix into the legacy service document.  Also cleaned
up some of the href generations for v2 to use the actual v2 helpers.
Fixed some testing issues as well.