🎁 Enhance the HykuCrosswalk to handle work create
This commit will create a V2 version of the HykuCrosswalk so it can be
used for create. Now we can generate the metadata of the record in the
response.  The root element changed from a `feed` to `entry` and made
sure that the work itself is the `link rel="edit"` and its file sets are
`link rel="edit-media` with the correct `href` attributes.  The content
element now has a `type="text/html"` as well and `href` changed to
`src`.  The properties now show in the response similar to a GET
request.