🧹 Atom compliant entry documents
This commit will bring entry documents in much closer compliance with
Atom specifications.  Also refactored the builder to pull some of that
logic out into the controller instead as well as getting rid of the
logic for V2 to check if it's Hyku because it will be Hyku.  We can
extract this out later into a configuration so this gem isn't so Hyku
focused, but for now this will do.