🐛 Fix HykuCrosswalk term mappings
Previously if given a term that was in the term_translation_mappings
that is supposed to be used for DC terms, HykuCrosswalk would transform
it when it shouldn't to build the metadata to ingets.  This commit will
fix that an align it with the DcCrosswalk.  This will fix the problem
with setting visibility as well.