🎁 Check if user is in the tenant
This commit will refactor the authorize_request method and also do an
additional check to see if the user is in the tenant.  This is a very
Hyku centric concept so if we ever want to make this gem work for Hyrax
we would need to rework the #allowed_access? method and override it in
Hyku to add the .for_repository scope.