🧹 Make POST for fileset inherit visibility
This commit will make is so that when a fileset gets created on a work,
it will inherit the visibility of its work instead of defaulting to
restricted.  At the moment, there is a limitation where even if you
supply the visibility of the fileset in the POST, it will still inherit
the visibility of the work.  Also, I had to update the rails helper
because the location of simple_work.rb changed in Hyrax it seems.  The
zip update was because I couldn't unzip it locally so I created a new
one.  Also we need to run the web service with `platform: linux/amd64`.